Render Mason quote to page.
Modified getEmoji() to return a emoji description also because, for some
reason, the passed through emoji wouldn't NOT be recogonized as the same
emoji. Believe it has something to do with the emoji being found in
another function (things seemed to work if the emoji variable is global).

This is what would happen in getMasonQuote():
    console.log(emoji); //returns "☁️"
    console.log(emoji === "☁️"); //returns "false"
idk...

Solution was to go back to getEmoji() and return an "emoji descriptor"
that can be use to look up the quote instead of the actual emoji.

Needs styling.